# octbase-api/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Octbase API
  version: "0.4.0"
  description: |
    REST API for Octbase — a focused project-management tool (tasks, boards,
    releases, sprints, wiki pages, SCM integration) for a single, invitation-only
    team.

    **Authentication:** All routes require a JWT **Bearer** access token in the
    `Authorization` header, except the public auth/invitation routes, the
    HMAC-authenticated webhook receivers, and the unauthenticated ops endpoints
    (`/api/v1/health`, `/api/v1/version`, `/api/v1/meta/enums`, `/api/v1/config`, `/metrics`,
    `/openapi.yaml`, `/docs`). Obtain an access token from `POST /api/v1/auth/login`;
    it is returned as `{ "accessToken", "expiresIn" }`. A long-lived refresh token is
    set as an HttpOnly cookie and rotated on every `POST /api/v1/auth/refresh`.

    **Authorization:** Two layers, both enforced server-side. A global role
    (`SUPER_ADMIN`, `ADMIN`, `USER`, `GUEST`) and a per-project membership role
    (`PROJECT_OWNER`, `PROJECT_ADMIN`, `PROJECT_MEMBER`, `PROJECT_VIEWER`). The
    effective project permission matrix is available at
    `GET /api/v1/projects/{projectId}/permissions`.

    **Error format:** All errors return
    `{ "code": "ERROR_CODE", "message": "...", "messageKey": "...", "details": {...} }`
    (`messageKey`/`details` optional).

    **Pagination:** List endpoints accept `?page=0&size=20` (zero-indexed pages,
    default size 20).

...
  contact:
    name: Octbase
...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    projectId: { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Conflict / version mismatch, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: Business-rule violation (e.g. LAST_OWNER, RELEASE_HAS_OPEN_TASKS), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        messageKey: { type: string }
        details: { type: object, additionalProperties: true }
...
    Sprint:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        name: { type: string }
        goal: { type: string }
        startDate: { type: string, format: date, nullable: true }
        endDate: { type: string, format: date, nullable: true }
        status: { type: string, enum: [PLANNED, ACTIVE, COMPLETED] }
        releaseId: { type: string, format: uuid, nullable: true }
        # Board scope of the sprint. For PLANNED/ACTIVE sprints these are filled
        # live from current sprint-board membership; for a COMPLETED sprint they
        # hold the totals snapshotted at completion (e.g. 2 of 5 done). Server-
        # computed, ignored on write.
        committedCount: { type: integer, readOnly: true }
        completedCount: { type: integer, readOnly: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer }
...
    SprintUpdate:
      description: >-
        Fields PATCH /sprints/{sprintId} actually applies. Any other key is
        rejected with 400 UNSUPPORTED_FIELD — status in particular transitions
        only via POST /sprints/{sprintId}/start and /complete.
      type: object
      properties:
        name: { type: string }
        goal: { type: string }
        startDate: { type: string, format: date, nullable: true }
        endDate: { type: string, format: date, nullable: true }
        releaseId: { type: string, format: uuid, nullable: true }
        version: { type: integer, description: 'The version this edit is based on; a stale value answers 409 VERSION_CONFLICT.' }
...
paths:
...
  /api/v1/projects/{projectId}/sprints:
    parameters: [ { $ref: '#/components/parameters/projectId' } ]
    get: { tags: [Planning], summary: List sprints, operationId: listSprints, responses: { '200': { description: Sprints, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Sprint' } } } } } } }
    post:
      tags: [Planning]
      summary: Create a sprint
      operationId: createSprint
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string }, goal: { type: string }, startDate: { type: string, format: date }, endDate: { type: string, format: date }, releaseId: { type: string, format: uuid, nullable: true } } } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Sprint' } } } }, '422': { $ref: '#/components/responses/Unprocessable' } }
  /api/v1/sprints/{sprintId}:
    parameters: [ { name: sprintId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get: { tags: [Planning], summary: Get a sprint, operationId: getSprint, responses: { '200': { description: Sprint, content: { application/json: { schema: { $ref: '#/components/schemas/Sprint' } } } }, '404': { $ref: '#/components/responses/NotFound' } } }
    patch:
      tags: [Planning]
      summary: Update a sprint
      operationId: updateSprint
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SprintUpdate' } } } }
      responses: { '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Sprint' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/Unprocessable' } }
    delete: { tags: [Planning], summary: Delete a sprint, operationId: deleteSprint, responses: { '204': { description: Deleted }, '404': { $ref: '#/components/responses/NotFound' } } }
  /api/v1/sprints/{sprintId}/start:
    parameters: [ { name: sprintId, in: path, required: true, schema: { type: string, format: uuid } } ]
    post:
      tags: [Planning]
      summary: Start a sprint (only one ACTIVE per project)
      operationId: startSprint
      responses: { '200': { description: Started, content: { application/json: { schema: { $ref: '#/components/schemas/Sprint' } } } }, '404': { $ref: '#/components/responses/NotFound' }, '422': { $ref: '#/components/responses/Unprocessable' } }
  /api/v1/sprints/{sprintId}/complete:
    parameters: [ { name: sprintId, in: path, required: true, schema: { type: string, format: uuid } } ]
    post:
      tags: [Planning]
      summary: Complete a sprint (unfinished tasks return to the backlog)
      operationId: completeSprint
      responses: { '200': { description: Completed, content: { application/json: { schema: { $ref: '#/components/schemas/Sprint' } } } }, '404': { $ref: '#/components/responses/NotFound' }, '422': { $ref: '#/components/responses/Unprocessable' } }

  # ───────────────────────── Reports ─────────────────────────
...